IssueReporter.jl Documentation

IssueReporter.jl Documentation

Functions

Index

Functions

IssueReporter.packageuri — Method

packageuri(pkgname)

Takes the name of a registered Julia package and returns the associated repo git URL. Examples

julia> IssueReporter.packageuri("IssueReporter")
"git://github.com/essenciary/IssueReporter.jl.git"

IssueReporter.tokenisdefined — Method

tokenisdefined()

Checks if the required GitHub authentication token is defined.

IssueReporter.token — Method

token()

Returns the configured GitHub authentication token, if defined – or throws an error otherwise.

IssueReporter.githubauth — Method

githubauth()

Performs GitHub authentication and returns the OAuth2 object, required by further GitHub API calls.

IssueReporter.repoid — Method

repoid(package_name)

Converts a registered Julia package name to the corresponding GitHub "username/repo_name" string. Examples

julia> IssueReporter.repoid("IssueReporter")
"essenciary/IssueReporter.jl"

IssueReporter.report — Method

report(package_name, title, body)

Creates a new GitHub issue with the title title and the content body onto the repo corresponding to the registered package called pack-age_name.

Index

IssueReporter.githubauth

IssueReporter.packageuri

IssueReporter.repoid

IssueReporter.report

IssueReporter.token

IssueReporter.tokenisdefined